+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## Current (in progress)
 
+- Allow to update static resources' files [#4](https://github.com/datagouv/datagouv_client/pull/4)
 - Polishing after first release [#2](https://github.com/datagouv/datagouv_client/pull/2)
 - Package creation [#1](https://github.com/datagouv/datagouv_client/pull/1)